Move Auth to it's own class
