+++ src/Feature/Auth/Types.hs
@@ -1,0 +1,24 @@
+module Feature.Auth.Types where
+
+import ClassyPrelude
+import Data.Aeson.TH
+
+type Token = Text
+type UserId = Integer
+type CurrentUser = (Token, UserId)
+
+data TokenError
+  = TokenErrorUserIdNotFound
+  | TokenErrorNotFound
+  | TokenErrorExpired
+  | TokenErrorMalformed String
+  deriving (Eq, Show)
+
+$(concat <$> 
+  mapM (\name -> 
+    let lowerCaseFirst (y:ys) = toLower [y] <> ys 
+        lowerCaseFirst "" = ""
+        structName = fromMaybe "" . lastMay . splitElem '.' . show $ name
+    in deriveJSON defaultOptions{fieldLabelModifier = lowerCaseFirst . drop (length structName)} name)
+  [ ''TokenError
+  ])
